Refactored examples from lessons into separate app
