+++ part1/lessonexamples/src/index.js
@@ -1,0 +1,26 @@
+import React, { useState } from 'react';
+import ReactDOM from 'react-dom';
+
+const Display = ({counter }) => <div>{counter}</div>
+
+const Button = ({ handleClick, text }) => (
+    <button onClick={handleClick}>
+      {text}
+    </button>
+)
+
+const App = () => {
+  const [ counter, setCounter ] = useState(0)
+  const increaseByOne = () => setCounter(counter + 1)
+  const setToZero = () => setCounter(0)
+
+  return (
+    <div>
+      <Display counter={counter} />
+      <Button handleClick={increaseByOne} text='plus' />
+      <Button handleClick={setToZero} text='reset' />
+    </div>
+  )
+}
+
+ReactDOM.render(<App />, document.getElementById('root'))
